// src/redux/pricing/initial-type.ts
export interface InitialType {
  [name: string]: any;
  brand: string;
  model: string;
  type: string;
  brand_model_error: boolean;
  year_of_manufacture: string;
  year: number;
  mileage: string;
  year_type: boolean;
  is_car_made_aboard: null;
  tab: number;
  estimate: number;
  lower_estimate: number;
  upper_estimate: number;
}